Update python.el
Please remove the needless clobbering of default settings for tab-width and indent-tabs-mode (lines 54800 and 5481).

If you Google the issue, there are literally thousands of people asking how to "fix" the tabs/ndentation in python-mode because it is one of the only programming language modes that clobbers rather than inherits default settings silently.

As currently implemented, using the previously used (setq python-indent-offset 4) or (setq-default python-indent-offset 4) still sprinkled around the internet as a "fix" does not work. Additionally, modifying defaults in the python-mode-hook does nothing due to how this is implemented:

(add-hook 'python-mode-hook
          (lambda ()
            (setq-default indent-tabs-mode t)
            (setq-default tab-width 4)))

Only

(add-hook 'python-mode-hook
          (lambda ()
            (setq indent-tabs-mode t)
            (setq tab-width 4)))

Seems to work.

A search of *every* other progmode shows that no other mode clobbers tab-width and only fortran, idlwave, vera-mode, and vhdl-mode touch 'intent-tabs-mode. Some of those have strict requirements for compilers, so it makes sense, Python does not.

If there is a compelling reason to try to enforce a particular indentation style, I would like to understand.